Automatically rotate tokens (#3)
* Adding always reset token setting support (by [tpb-dev])

* Change token reset from per-request to per-login

Tokens shouldn't be rotated on every single request, because that would
mean each request would require a new "login" request to get the new
token. Instead, we want to reset the token every time login credentials
are sent to get the token that should be used for authentication. This
is the desired behavior - when someone logs in, revoke any other tokens
regardless of expiration. This ensures that only one client is logged in
at once.

* [minor] formatting fix